requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv==1.0.0